Add files via upload
This project allows users to guess a randomly generated number between 1 and 100. The program provides hints such as "Too high!" or "Too low!" to guide the user towards the correct number. The goal is to guess the number within the fewest attempts possible. The project tracks the number of guesses and displays a congratulatory message when the correct number is guessed.